// rk-cb75-source/keyboard/beiying/enum.ts
export enum KeyMatrixLayer {
    Nomal = 0x00,
    FN1 = 0x01,
    FN2 = 0x02,
    Tap = 0x03
}

export enum MatrixTable {
    WIN = 0x00,
    MAC = 0x01
}

export enum LightEffectEnum {
    FixedOn = 1,
    Respire = 2,
    Rainbow = 3,
    FlashAway = 4,
    Raindrops = 5,
    RainbowWheel = 6,
    RippleShining = 7,
    StarsTwinkle = 8,
    ShadowDisappear = 9,
    RetroSnake = 10,
    NeonStream = 11,
    Reaction = 12,
    SineWave = 13,
    SideScan = 14,
    RotatingWindmill = 15,
    ColorfulWaterfall = 16,
    Blossoming = 17,
    SelfDefine = 18,
    OFF = 0,
    Music = 19
}